feat: Add support for `#[sv::]` attributes for all sylvia attribtues. (#310)
This PR adds the requirement for `sv::` prefix for all attributes
handled by sylvia. Attributes wihtout `sv::` prefix are not supported
and are not handled by the macro.

Other than that, this change also introduces refactoring commits - the
parser has been split into a multi-file module, and the interface has
been unified.